refactor: move gen-schedule-helper into directory where its belong
