+++ app/src/functions/gen-schedule-helper.ts
